Make directories analysis-specific
An analysis is defined as the following: an outcome of interest,
a set of structured data features and silver labels, and
a set of decisions (including whether to use negation in computing
NLP variables, whether to include normalized versions of NLP variables,
whether to perform dimension reduction). Directory names are now
analysis-specific to avoid overwriting output.