Critical updates to getting predicted probabilities
First, add random number seed as argument; this ensures that
the same random number seed will be used to generate predicted
probabilities on the training dataset as was used for the PheNorm
fitting. Second, ensure that the study ID column is only removed
from the training data if it is present (which it should not be,
since preprocessing removes the ID).